ci: add GitHub Actions build workflow
A new build workflow has been added to the .github/workflows directory. This workflow is set to trigger on push events to master, bugfix, hotfix, and release branches, except for markdown file changes. It also triggers on pull request events and allows manual running from the Actions tab.

The workflow is composed of three jobs named 'guard', 'build' and 'release', which are responsible for running the usage guard, building and testing the code, and creating a GitHub release respectively. The jobs also make use of specific actions to speed up builds, install toolchains, check for errors, generate a changelog, set the release info, and create the release. Furthermore, the workflow has a concurrency setting to control the concurrency level of the job and cancel previous runs in progress for the same group-key.