chore: add new GraphQL schema for communities and platforms
This commit adds a new GraphQL schema to represent the data of NounsDAO - a decentralized autonomous organization built on Ethereum blockchain. The schema represents entities of the NounsDAO such as Account, Noun, Delegate, Vote and Proposal along with directives for querying and manipulating these entities.

For example, there are Account type and Account_filter input to filter accounts based on various fields. Similarly, there are types to represent Noun, Delegate and Proposal along with their properties.

It aims to provide a structured way to query and manipulate data of NounsDAO in a structured way from a GraphQL client.