refactor: modify font style in `create_og_image` function
The `create_og_image` function in `utils.rs` has been modified to refactor the font style. The function previously used the `Londrina Solid` font for both the title and the meta, but now it separates them and uses different font sizes. The title is now displayed in the `Londrina Solid_60_bold` font and the meta is displayed in the `Londrina Solid_40` font. These changes improve the visual appearance of the generated image.