ci: update deployment command in Github Action
Command to deploy to 'dev' environment has been updated to be more generic in the 'deploy.yml' Github Action workflow. With this update, the script will now fetch environment variables dynamically, rather than hard-coding the 'dev' environment. This change is aimed to make the deployment process more flexible and adaptable to changes in environments.