chore: update `deploy` script in `package.json`
The deploy script in the package.json file has been updated with an added "--env dev" flag. This change has been made to ensure that the application is always deployed to the development environment. This will help in maintaining the consistency across multiple stages of the pipeline and ensure a smoother and more reliable deployment process.